Add support for copilot globs instructions
