+++ .github/instructions/protobuf.instructions.md
@@ -1,7 +1,5 @@
 ---
-description:
-globs: **/*.proto
-alwaysApply: false
+applyTo: "**/*.proto"
 ---
 
 # Protobuf
