+++ .github/instructions/python_lib.instructions.md
@@ -1,0 +1,35 @@
+---
+applyTo: "lib/streamlit/**/*.py"
+---
+
+# Streamlit Lib Python Guide
+
+Tips and guidelines specific to the development of the Streamlit Python library,
+not applicable scripts and e2e tests.
+
+## Logging
+
+If something needs to be logged, please use our logger - that returns a default
+Python logger - with an appropriate logging level:
+
+```python
+from streamlit.logger import get_logger
+
+_LOGGER: Final = get_logger(__name__)
+```
+
+## Unit Tests
+
+We use the unit tests to cover internal behavior that can work without the web / backend
+counterpart and the e2e tests to test the entire system. We aim for high unit test
+coverage (90% or higher) of our Python code in `lib/streamlit`.
+
+- Under `lib/tests/streamlit`, add a new test file
+- Preferably in the mirrored directory structure as the non-test files.
+- Naming: `my_example_test.py`
+
+### Typing Tests
+
+We also have typing tests in `lib/tests/streamlit/typing` for our public API to catch
+typing errors in parameters or return types by using mypy and `assert_type`.
+Check other typing tests in the `lib/tests/streamlit/typing` directory for inspiration.
