+++ .pre-commit-config.yaml
@@ -96,18 +96,16 @@ repos:
         language: system
         always_run: true
         pass_filenames: false
-      - id: generate-cursor-rules
-        name: Check generated cursor rules
-        entry: ./scripts/generate_cursor_rules.py
+      - id: generate-agent-rules
+        name: Check generated agent rules
+        entry: ./scripts/generate_agent_rules.py
         language: system
-        files: ^Makefile$
         pass_filenames: false
-      - id: generate-copilot-instructions
-        name: Check generated copilot instructions
-        entry: ./scripts/generate_copilot_instructions.py
-        language: system
-        files: ^.cursor/.*\.(mdc)$
-        pass_filenames: false
+        files: |
+          (?x)
+          (^|.*/)AGENTS\.md$
+          |^scripts/generate_agent_rules\.py$
+          |^Makefile$
       - id: vscode-devcontainer-sync
         name: Check VSCode/devcontainer sync
         entry: ./scripts/sync_vscode_devcontainer.py --check
